+++ rust/operator-binary/src/controller.rs
@@ -7,14 +7,16 @@ use crate::product_logging::{
 };
 
 use snafu::{OptionExt, ResultExt, Snafu};
-use stackable_opa_crd::{Container, OpaCluster, OpaConfig, OpaRole, APP_NAME, OPERATOR_NAME};
-use stackable_operator::builder::{PodSecurityContextBuilder, VolumeBuilder};
-use stackable_operator::commons::rbac::build_rbac_resources;
-use stackable_operator::kube::ResourceExt;
-use stackable_operator::product_logging::spec::{AppenderConfig, LogLevel};
+use stackable_opa_crd::{
+    Container, OpaCluster, OpaClusterStatus, OpaConfig, OpaRole, APP_NAME, OPERATOR_NAME,
+};
 use stackable_operator::{
-    builder::{ConfigMapBuilder, ContainerBuilder, FieldPathEnvVar, ObjectMetaBuilder, PodBuilder},
-    commons::product_image_selection::ResolvedProductImage,
+    builder::{
+        ConfigMapBuilder, ContainerBuilder, FieldPathEnvVar, ObjectMetaBuilder, PodBuilder,
+        PodSecurityContextBuilder, VolumeBuilder,
+    },
+    cluster_resources::{ClusterResourceApplyStrategy, ClusterResources},
+    commons::{product_image_selection::ResolvedProductImage, rbac::build_rbac_resources},
     k8s_openapi::{
         api::{
             apps::v1::{DaemonSet, DaemonSetSpec},
@@ -26,16 +28,26 @@ use stackable_operator::{
             api::resource::Quantity, apis::meta::v1::LabelSelector, util::intstr::IntOrString,
         },
     },
-    kube::runtime::{controller::Action, reflector::ObjectRef},
+    kube::{
+        runtime::{controller::Action, reflector::ObjectRef},
+        Resource as KubeResource, ResourceExt,
+    },
     labels::{role_group_selector_labels, role_selector_labels, ObjectLabels},
     logging::controller::ReconcilerError,
     product_config::{types::PropertyNameKind, ProductConfigManager},
     product_config_utils::{transform_all_roles_to_config, validate_all_roles_and_groups_config},
     product_logging::{
         self,
-        spec::{AutomaticContainerLogConfig, ContainerLogConfig, ContainerLogConfigChoice},
+        spec::{
+            AppenderConfig, AutomaticContainerLogConfig, ContainerLogConfig,
+            ContainerLogConfigChoice, LogLevel,
+        },
     },
     role_utils::RoleGroupRef,
+    status::condition::{
+        compute_conditions, daemonset::DaemonSetConditionBuilder,
+        operations::ClusterOperationsConditionBuilder,
+    },
 };
 use std::{
     borrow::Cow,
@@ -98,10 +110,6 @@ pub enum Error {
     ApplyRoleService {
         source: stackable_operator::error::Error,
     },
-    #[snafu(display("failed to apply global RoleBinding"))]
-    ApplyRoleRoleBinding {
-        source: stackable_operator::error::Error,
-    },
     #[snafu(display("failed to apply Service for {}", rolegroup))]
     ApplyRoleGroupService {
         source: stackable_operator::error::Error,
@@ -121,6 +129,20 @@ pub enum Error {
     ApplyRoleGroupDaemonSet {
         source: stackable_operator::error::Error,
         rolegroup: RoleGroupRef<OpaCluster>,
+    },
+    #[snafu(display("failed to patch service account: {source}"))]
+    ApplyServiceAccount {
+        name: String,
+        source: stackable_operator::error::Error,
+    },
+    #[snafu(display("failed to patch role binding: {source}"))]
+    ApplyRoleBinding {
+        name: String,
+        source: stackable_operator::error::Error,
+    },
+    #[snafu(display("failed to update status"))]
+    ApplyStatus {
+        source: stackable_operator::error::Error,
     },
     #[snafu(display("invalid product config"))]
     InvalidProductConfig {
@@ -155,14 +177,12 @@ pub enum Error {
         source: crate::product_logging::Error,
         cm_name: String,
     },
-    #[snafu(display("failed to patch service account: {source}"))]
-    ApplyServiceAccount {
-        name: String,
+    #[snafu(display("failed to create cluster resources"))]
+    FailedToCreateClusterResources {
         source: stackable_operator::error::Error,
     },
-    #[snafu(display("failed to patch role binding: {source}"))]
-    ApplyRoleBinding {
-        name: String,
+    #[snafu(display("failed to delete orphaned resources"))]
+    DeleteOrphans {
         source: stackable_operator::error::Error,
     },
 }
@@ -177,8 +197,17 @@ impl ReconcilerError for Error {
 pub async fn reconcile_opa(opa: Arc<OpaCluster>, ctx: Arc<Ctx>) -> Result<Action> {
     tracing::info!("Starting reconcile");
     let opa_ref = ObjectRef::from_obj(opa.as_ref());
-    let client = ctx.client.clone();
+    let client = &ctx.client;
     let resolved_product_image = opa.spec.image.resolve(DOCKER_IMAGE_BASE_NAME);
+
+    let mut cluster_resources = ClusterResources::new(
+        APP_NAME,
+        OPERATOR_NAME,
+        OPA_CONTROLLER_NAME,
+        &opa.object_ref(&()),
+        ClusterResourceApplyStrategy::from(&opa.spec.cluster_operation),
+    )
+    .context(FailedToCreateClusterResourcesSnafu)?;
 
     let (rbac_sa, rbac_rolebinding) = build_rbac_resources(opa.as_ref(), "opa");
     client
@@ -222,18 +251,17 @@ pub async fn reconcile_opa(opa: Arc<OpaCluster>, ctx: Arc<Ctx>) -> Result<Action
         .unwrap_or_default();
 
     let server_role_service = build_server_role_service(&opa, &resolved_product_image)?;
-    let server_role_service = client
-        .apply_patch(
-            OPA_CONTROLLER_NAME,
-            &server_role_service,
-            &server_role_service,
-        )
+    // required for discovery config map later
+    let server_role_service = cluster_resources
+        .add(client, server_role_service)
         .await
         .context(ApplyRoleServiceSnafu)?;
 
     let vector_aggregator_address = resolve_vector_aggregator_address(&opa, &client)
         .await
         .context(ResolveVectorAggregatorAddressSnafu)?;
+
+    let mut ds_cond_builder = DaemonSetConditionBuilder::default();
 
     for (rolegroup_name, rolegroup_config) in role_server_config.iter() {
         let rolegroup = RoleGroupRef {
@@ -253,6 +281,7 @@ pub async fn reconcile_opa(opa: Arc<OpaCluster>, ctx: Arc<Ctx>) -> Result<Action
             &merged_config,
             vector_aggregator_address.as_deref(),
         )?;
+        let rg_service = build_rolegroup_service(&opa, &resolved_product_image, &rolegroup)?;
         let rg_daemonset = build_server_rolegroup_daemonset(
             &opa,
             &resolved_product_image,
@@ -261,26 +290,27 @@ pub async fn reconcile_opa(opa: Arc<OpaCluster>, ctx: Arc<Ctx>) -> Result<Action
             &merged_config,
             &rbac_sa.name_unchecked(),
         )?;
-        let rg_service = build_rolegroup_service(&opa, &resolved_product_image, &rolegroup)?;
 
-        client
-            .apply_patch(OPA_CONTROLLER_NAME, &rg_configmap, &rg_configmap)
+        cluster_resources
+            .add(client, rg_configmap)
             .await
             .with_context(|_| ApplyRoleGroupConfigSnafu {
                 rolegroup: rolegroup.clone(),
             })?;
-        client
-            .apply_patch(OPA_CONTROLLER_NAME, &rg_daemonset, &rg_daemonset)
-            .await
-            .with_context(|_| ApplyRoleGroupDaemonSetSnafu {
-                rolegroup: rolegroup.clone(),
-            })?;
-        client
-            .apply_patch(OPA_CONTROLLER_NAME, &rg_service, &rg_service)
+        cluster_resources
+            .add(client, rg_service)
             .await
             .with_context(|_| ApplyRoleGroupServiceSnafu {
                 rolegroup: rolegroup.clone(),
             })?;
+        ds_cond_builder.add(
+            cluster_resources
+                .add(client, rg_daemonset)
+                .await
+                .with_context(|_| ApplyRoleGroupDaemonSetSnafu {
+                    rolegroup: rolegroup.clone(),
+                })?,
+        );
     }
 
     for discovery_cm in build_discovery_configmaps(
@@ -291,11 +321,31 @@ pub async fn reconcile_opa(opa: Arc<OpaCluster>, ctx: Arc<Ctx>) -> Result<Action
     )
     .context(BuildDiscoveryConfigSnafu)?
     {
-        client
-            .apply_patch(OPA_CONTROLLER_NAME, &discovery_cm, &discovery_cm)
+        cluster_resources
+            .add(client, discovery_cm)
             .await
             .context(ApplyDiscoveryConfigSnafu)?;
     }
+
+    let cluster_operation_cond_builder =
+        ClusterOperationsConditionBuilder::new(&opa.spec.cluster_operation);
+
+    let status = OpaClusterStatus {
+        conditions: compute_conditions(
+            opa.as_ref(),
+            &[&ds_cond_builder, &cluster_operation_cond_builder],
+        ),
+    };
+
+    client
+        .apply_patch_status(OPERATOR_NAME, &*opa, &status)
+        .await
+        .context(ApplyStatusSnafu)?;
+
+    cluster_resources
+        .delete_orphaned_resources(client)
+        .await
+        .context(DeleteOrphansSnafu)?;
 
     Ok(Action::await_change())
 }
@@ -567,6 +617,7 @@ fn build_server_rolegroup_daemonset(
     .add_container(cb_opa.build())
     .add_container(cb_bundle_builder.build())
     .image_pull_secrets_from_product_image(resolved_product_image)
+    .node_selector_opt(opa.node_selector(&rolegroup_ref.role_group))
     .add_volume(
         VolumeBuilder::new(CONFIG_VOLUME_NAME)
             .with_config_map(rolegroup_ref.object_name())